TCC730 slight stylistic improvement
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@6081 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/export/system.h
@@ -235,7 +235,7 @@ static inline unsigned long SWAB32(unsigned long value)
       result[ 7.. 0] = value[31..24];
     */
 {
-    unsigned short hi = SWAB16(value >> 16);
+    unsigned long hi = SWAB16(value >> 16);
     unsigned long lo = SWAB16(value & 0xffff);
     return (lo << 16) | hi;
 }
